ada: Add special-case for 'Constrained on stand-alone objects in SPARK
GNAT relies on standalone objects of indefinite, non-class-wide types getting a
constrained itypes; GNATprove relies on the same objects keeping their nominal,
unconstrained types. None of that can be simply changed, so instead we carry
this different handling to routine which provides the value of attribute
Constrained.

gcc/ada/ChangeLog:

	* exp_util.adb (Attribute_Constrained_Static_Value): Special case
	stand-alone objects for GNATprove.