tree-optimization/120924 - up --param uninit-max-chain-len
The PR shows that the uninit analysis limits are set too low in
cases we lower switches to ifs as happens on s390x for a linux
kernel TU.  This causes false positive uninit diagnostics as we
abort the attempt to prove that a value is initialized on all
paths.  The new testcase only would require upping to 9.

	PR tree-optimization/120924
	* params.opt (uninit-max-chain-len): Up from 8 to 12.

	* gcc.dg/uninit-pr120924.c: New testcase.