RISC-V: Add ashiftrt operand 2 for vector avg_floor and avg_ceil
According to the semantics of the avg_floor and avg_ceil as below:

floor: op0 = (narrow) (((wide) op1 + (wide) op2) >> 1);
ceil:  op0 = (narrow) (((wide) op1 + (wide) op2 + 1) >> 1);

Aka we have (const_int 1) as the op2 of the ashiftrt but seems missed.
Thus, add it back to align the definition.

The below test suites are passed for this patch.
* The rv64gcv fully regression test.

gcc/ChangeLog:

	* config/riscv/autovec.md: Add (const_int 1) as the op2 of
	ashiftrt.

Signed-off-by: Pan Li <pan2.li@intel.com>