RISC-V: Prepare dynamic LMUL heuristic for SLP.
This patch prepares the dynamic LMUL vector costing to use the coming
SLP_TREE_TYPE instead of the (to-be-removed) STMT_VINFO_TYPE.

Even though the whole approach should be reviewed and adjusted at some
point, the patch chooses the path of least resistance and uses a hash
map for the stmt_info -> slp node relationship.  A node is mapped to the
accompanying stmt_info during add_stmt_cost.  In finish_cost we go
through all statements as before, and obtain the corresponding slp nodes
as well as their types.

This allows us to operate largely as before.  We don't yet do the switch
over from STMT_VINFO_TYPE to SLP_TREE_TYPE, though but only take care
of the necessary refactoring upfront.

Regtested on rv64gcv_zvl512b with -mrvv-max-lmul=dynamic.  There are a
few regressions but nothing worse than what we already have.  I'd rather
accept these now and take it as an incentive to work on the heuristic
later than block the SLP work until it is fixed.

gcc/ChangeLog:

	* config/riscv/riscv-vector-costs.cc (get_live_range):
	Move compute_local_program_points to cost class.
	(variable_vectorized_p): Add slp node parameter.
	(need_additional_vector_vars_p): Move from here...
	(costs::need_additional_vector_vars_p): ... to here and add slp
	parameter.
	(compute_estimated_lmul): Move update_local_live_ranges to cost
	class.
	(has_unexpected_spills_p): Move from here...
	(costs::has_unexpected_spills_p): ... to here.
	(costs::record_lmul_spills): New function.
	(costs::add_stmt_cost): Add stmt_info, slp mapping.
	(costs::finish_cost): Analyze loop.
	* config/riscv/riscv-vector-costs.h: Move declarations to class.