tree-optimization/121202 - fix vector stmt placement
When we have a vector shift with a scalar the shift operand can be
external - in that case we should not use the shift operand def
as hint where to place the vector shift instruction.  The ICE
in the PR is because stmt dominance queries only work inside of
the vector region.  But we should also never place stmts outside
of it.

	PR tree-optimization/121202
	* tree-vect-slp.cc (vect_schedule_slp_node): Do not take
	an out-of-region stmt as "last".

	* gcc.dg/pr121202.c: New testcase.