tree-optimization/121220 - improve sinking of stores
We currently do only very restricted store sinking into paths
that have no loads or stores and end in a virtual PHI.  The
following extends this to sink towards a single virtual
definition in addition to the case of a PHI, handling skipping
of unrelated virtual uses.  We later have to prune cases
that would require virtual PHI insertion and the patch below
basically restricts this to sinking to noreturn paths for now.

	PR tree-optimization/121220
	* tree-ssa-sink.cc (statement_sink_location): For stores
	handle sinking to paths ending in a store.  Skip loads
	that do not use the store.

	* gcc.dg/tree-ssa/ssa-sink-23.c: New testcase.