Don't pass vector params through to offload targets
The optimization options are deliberately passed through to the LTO compiler,
but when the same mechanism is reused for offloading it ends up forcing the
host compiler settings onto the device compiler.  Maybe this should be removed
completely, but this patch just fixes a few of them.  In particular,
param_vect_partial_vector_usage is disabled by x86 and this really hurts amdgcn.

I also fixed an ambiguous else warning in the generated file by adding braces.

gcc/ChangeLog:

	* config/gcn/gcn.cc (gcn_option_override): Add note to set default for
	param_vect_partial_vector_usage to "1".
	* optc-save-gen.awk: Don't pass through options marked "NoOffload".
	* params.opt (-param=vect-epilogues-nomask): Add NoOffload.
	(-param=vect-partial-vector-usage): Likewise.
	(-param=vect-inner-loop-cost-factor): Likewise.