Remove now redundant vect_get_vec_defs overload
The following removes the vect_get_vec_defs overload receiving
a vector type to be used for the possibly constant/invariant
operand.  This was used for non-SLP code generation as there
constants/invariants are generated on the fly.  It also elides
the stmt_vec_info and ncopies argument which are not required
for SLP.

	* tree-vectorizer.h (vect_get_vec_defs): Remove overload
	with operand vector type.  Remove stmt_vec_info and
	ncopies argument.
	* tree-vect-stmts.cc (vect_get_vec_defs): Likewise.
	(vectorizable_conversion): Adjust by not passing in
	vector types, stmt_vec_info and ncopies.
	(vectorizable_bswap): Likewise.
	(vectorizable_assignment): Likewise.
	(vectorizable_shift): Likewise.
	(vectorizable_operation): Likewise.
	(vectorizable_scan_store): Likewise.
	(vectorizable_store): Likewise.
	(vectorizable_condition): Likewise.
	(vectorizable_comparison_1): Likewise.
	* tree-vect-loop.cc (vect_transform_reduction): Likewise.
	(vect_transform_lc_phi): Likewise.